Estimate Standardized Regression Coefficients and Generate the Corresponding Sampling Distribution Using the Monte Carlo Method

Estimate Standardized Regression Coefficients and Generate the Corresponding Sampling Distribution Using the Monte Carlo Method

Usage

BetaMC(object)

Arguments

object

Object of class mc, that is, the output of the MC() function.

Value

Returns an object of class betamc which is a list with the following elements:

object

The function argument object.

thetahatstar

Sampling distribution of \(\boldsymbol{\hat{\beta}}\).

vcov

Sampling variance-covariance matrix of \(\boldsymbol{\hat{\beta}}\).

est

Vector of estimated \(\boldsymbol{\hat{\beta}}\).

fun

Function used ("BetaMC").

Details

The vector of standardized regression coefficients (\(\boldsymbol{\hat{\beta}}\)) is derived from each randomly generated vector of parameter estimates. Confidence intervals are generated by obtaining percentiles corresponding to \(100(1 - \alpha)\%\) from the generated sampling distribution of \(\boldsymbol{\hat{\beta}}\), where \(\alpha\) is the significance level.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# Fit the regression model
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
# Generate the sampling distribution of parameter estimates
# (use a large R, for example, R = 20000 for actual research)
mc <- MC(object, R = 100)
# Generate confidence intervals for standardized regression slopes
std <- BetaMC(mc)
# Methods --------------------------------------------------------
print(std)
#> Standardized regression slopes
#> type = "hc3"
#>            est     se   R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> NARTIC  0.4951 0.0808 100 0.2798 0.2974 0.3272 0.6403 0.6686 0.6839
#> PCTGRT  0.3915 0.0845 100 0.1845 0.1884 0.2082 0.5303 0.5677 0.5852
#> PCTSUPP 0.2632 0.0848 100 0.0140 0.0577 0.1181 0.4265 0.4470 0.4569
summary(std)
#> Standardized regression slopes
#> type = "hc3"
#>            est     se   R  0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> NARTIC  0.4951 0.0808 100 0.2798 0.2974 0.3272 0.6403 0.6686 0.6839
#> PCTGRT  0.3915 0.0845 100 0.1845 0.1884 0.2082 0.5303 0.5677 0.5852
#> PCTSUPP 0.2632 0.0848 100 0.0140 0.0577 0.1181 0.4265 0.4470 0.4569
coef(std)
#>    NARTIC    PCTGRT   PCTSUPP 
#> 0.4951451 0.3914887 0.2632477 
vcov(std)
#>               NARTIC       PCTGRT      PCTSUPP
#> NARTIC   0.006533301 -0.003999164 -0.001252905
#> PCTGRT  -0.003999164  0.007132101 -0.003466663
#> PCTSUPP -0.001252905 -0.003466663  0.007190586
confint(std, level = 0.95)
#>              2.5%     97.5%
#> NARTIC  0.3271645 0.6402921
#> PCTGRT  0.2082088 0.5302910
#> PCTSUPP 0.1180726 0.4265075