Estimate Semipartial Correlation Coefficients and Generate the Corresponding Sampling Distribution Using the Monte Carlo Method

Estimate Semipartial Correlation Coefficients and Generate the Corresponding Sampling Distribution Using the Monte Carlo Method

Usage

SCorMC(object, alpha = c(0.05, 0.01, 0.001))

Arguments

object

Object of class mc, that is, the output of the MC() function.

alpha

Numeric vector. Significance level \(\alpha\).

Value

Returns an object of class betamc which is a list with the following elements:

call

Function call.

args

Function arguments.

thetahatstar

Sampling distribution of \(r_{s}\).

vcov

Sampling variance-covariance matrix of \(r_{s}\).

est

Vector of estimated \(r_{s}\).

fun

Function used ("SCorMC").

Details

The vector of semipartial correlation coefficients (\(r_{s}\)) is derived from each randomly generated vector of parameter estimates. Confidence intervals are generated by obtaining percentiles corresponding to \(100(1 - \alpha)\%\) from the generated sampling distribution of \(r_{s}\), where \(\alpha\) is the significance level.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# Data ---------------------------------------------------------------------
data("nas1982", package = "betaMC")

# Fit Model in lm ----------------------------------------------------------
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)

# MC -----------------------------------------------------------------------
mc <- MC(
  object,
  R = 100, # use a large value e.g., 20000L for actual research
  seed = 0508
)

# SCorMC -------------------------------------------------------------------
out <- SCorMC(mc, alpha = 0.05)

## Methods -----------------------------------------------------------------
print(out)
#> Call:
#> SCorMC(object = mc, alpha = 0.05)
#> 
#> Semipartial correlations
#> type = "hc3"
#>            est     se   R   2.5%  97.5%
#> NARTIC  0.4312 0.0874 100 0.2295 0.5623
#> PCTGRT  0.3430 0.0778 100 0.1650 0.4643
#> PCTSUPP 0.2385 0.0720 100 0.0880 0.3569
summary(out)
#> Call:
#> SCorMC(object = mc, alpha = 0.05)
#> 
#> Semipartial correlations
#> type = "hc3"
#>            est     se   R   2.5%  97.5%
#> NARTIC  0.4312 0.0874 100 0.2295 0.5623
#> PCTGRT  0.3430 0.0778 100 0.1650 0.4643
#> PCTSUPP 0.2385 0.0720 100 0.0880 0.3569
coef(out)
#>    NARTIC    PCTGRT   PCTSUPP 
#> 0.4311525 0.3430075 0.2384789 
vcov(out)
#>                NARTIC        PCTGRT       PCTSUPP
#> NARTIC   0.0076365444  0.0016690137 -0.0003859994
#> PCTGRT   0.0016690137  0.0060497385 -0.0006650079
#> PCTSUPP -0.0003859994 -0.0006650079  0.0051908878
confint(out, level = 0.95)
#>              2.5 %    97.5 %
#> NARTIC  0.22954565 0.5622638
#> PCTGRT  0.16503602 0.4642695
#> PCTSUPP 0.08800891 0.3569279