Estimated Parameter Method for an Object of Class betamc

Estimated Parameter Method for an Object of Class betamc

Usage

# S3 method for class 'betamc'
coef(object, ...)

Arguments

object

Object of Class betamc, that is, the output of the BetaMC(), RSqMC(), SCorMC(), DeltaRSqMC(), PCorMC(), or DiffBetaMC() functions.

...

additional arguments.

Value

Returns a vector of estimated parameters.

Author

Ivan Jacob Agaloos Pesigan